Type engine updated and unused code deleted
